Modifying long to integer

+++ src/main/groovy/br/com/mlm/dostock/domain/ProductBatch.groovy
@@ -1,10 +1,6 @@
 package br.com.mlm.dostock.domain
 
-import javax.persistence.Column
-import javax.persistence.Entity
-import javax.persistence.OneToOne
-import javax.persistence.Temporal
-import javax.persistence.TemporalType
+import javax.persistence.*
 
 @Entity
 class ProductBatch extends BaseEntity{
@@ -12,7 +8,7 @@ class ProductBatch extends BaseEntity{
     @Column(nullable = false)
     String number
 
-    Long quantity = 0
+    Integer quantity = 0
 
     @Temporal(TemporalType.TIMESTAMP)
     Date validate
